LayoutTests/media/media-source/media-managedmse-waiting-at-gap-expected.txt


RUN(video.src = URL.createObjectURL(source))
RUN(video.muted = true)
RUN(video.playsInline = true)
RUN(video.disableRemotePlayback = true)
RUN(video.autoplay = true)
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(1)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(2)))
EVENT(update)
EVENT(waiting)
EXPECTED (video.currentTime == video.buffered.end(0) == 'true') OK
EXPECTED (video.readyState == '2') OK
EXPECTED (readyStateAtLastTimeupdate == '2') OK
END OF TEST
